All Lockfile Refactoring
