+++ conda_lock/conda_solver.py
@@ -31,14 +31,9 @@ from conda_lock.invoke_conda import (
     conda_pkgs_dir,
     is_micromamba,
 )
+from conda_lock.lockfile import HashModel, LockedDependency, _apply_categories
 from conda_lock.models.channel import Channel
-from conda_lock.src_parser import (
-    Dependency,
-    HashModel,
-    LockedDependency,
-    VersionedDependency,
-    _apply_categories,
-)
+from conda_lock.src_parser import Dependency, VersionedDependency
 
 
 logger = logging.getLogger(__name__)
