+++ conda_lock/models/__init__.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel
+
+
+class StrictModel(BaseModel):
+    class Config:
+        extra = "forbid"
+        json_encoders = {
+            frozenset: list,
+        }
